% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{Check_Normal}
\alias{Check_Normal}
\title{Check Normality}
\usage{
Check_Normal(data, plot_type)
}
\arguments{
\item{data}{Data collected in tibble format that will be visualized and
tested for normality}

\item{plot_type}{Plot used to visualize data "histogram" or "QQ".}
}
\value{
Either a histogram or QQ plot and a Shapiro-Wilk test for normality.
}
\description{
Conducts a Shapiro-Wilk test for the control and experimental
columns as well as generating either a histogram or QQ plot so that students
can visualize their data distributions and quantify normality.
}
\examples{
Check_Normal(data, "histogram")
}
