% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{Create_Figure}
\alias{Create_Figure}
\title{Create Figure}
\usage{
Create_Figure(data, errorbars, test_level, yaxis_label, group)
}
\arguments{
\item{data}{Data collected in tibble format that will be used to create a
figure.}

\item{errorbars}{Error bars for control and experimental variable with
options of standard deviation "sd" or standard error "se"}

\item{test_level}{Either "parametric" or "nonparametric" determined when
using the Check_Normal function}

\item{yaxis_label}{A "string" that labels the y-axis. If no y-axis label is
desired then enter NULL}

\item{group}{A "string" that labels the control and experimental variables.
If the experiment tested different antibiotic treatments the the group would
be "Antibiotic Treatment". If no group label is desired then enter NULL.}
}
\value{
A figure with no theme that shows averages for parametric data or
medians for nonparametric data with error bars of sd or se.
}
\description{
A standardized figure that has no set theme. theme_set() from
ggplot2 can be used in order to customize the figure. theme_set(minimal) is a
common theme.
}
\examples{
Create_Figure(data, errorbars = "se", test_level = "parametric",
yaxis_label = "Average of an Outcome",
group = "Commonality of the Two Variables")
}
