% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{Outlier_Detect}
\alias{Outlier_Detect}
\title{Outlier Detect}
\usage{
Outlier_Detect(data, col_c, col_x, paired, REMOVE, plot_type)
}
\arguments{
\item{data}{Data collected in tibble format that will be analyzed for
outliers.}

\item{col_c}{The number of the column that contains the control variable
(as.numeric).}

\item{col_x}{The number of the column that contains the experimental variable
(as.numeric) .}

\item{paired}{Is the data is paired or not (T/F).}

\item{REMOVE}{Outliers should be removed from the data (T/F).}

\item{plot_type}{What type of plot to be created "box" for independent data
or "QQ" for paired data.}
}
\value{
A graphical depiction of the data with or without outliers.
}
\description{
Detects outliers in two columns of a tibble. The control and the
experimental. When assigned to a new object returns a tibble that has the
control variable in the first column (output_data[1)] and the experimental
variable in the second column (output_data[2]).

[1)]: R:1)
[2]: R:2
}
\examples{
output_data <- Outlier_Detect(data, col_c = 1, col_x = 2,
paired = F, REMOVE = F, plot_type = "box")
}
