% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{Test_Data}
\alias{Test_Data}
\title{Test Data}
\usage{
Test_Data(data, paired, test_level, input_alternative, mu_given)
}
\arguments{
\item{data}{Data collected in tibble format that will be tested for
significance.}

\item{paired}{Is the data is paired or not (T/F).}

\item{test_level}{Either "parametric" or "nonparametric" determined when
using the Check_Normal function}

\item{input_alternative}{The sidedness of test used for the alternative
hypothesis "less"/"two.sided"/"greater".}

\item{mu_given}{The null hypothesis of the given test. Most often the null
hypothesis is that the difference between means is 0.}
}
\value{
A p-value of a statistical test used based off the inputs into the
function.
}
\description{
Conducts various tests based off of user input. The test may be
parametric or nonparametric. It also takes into account if the variables are
paired, the alternative hypothesis "less", "two.sided", or "greater", and the
difference in mu which is typically zero.
}
\examples{
Test_Data(data, paired = F, test_level = "parametric",
input_alternative = "two.sided", mu_given = 0)
}
